delete other page
