+++ src/pages/index.js
@@ -1,151 +1,841 @@
-import React from 'react';
-import Button from '../components/Button';
-import Card from '../components/Card';
-import CustomerCard from '../components/CustomerCard';
-import LabelText from '../components/LabelText';
-import Layout from '../components/layout/Layout';
-import SplitSection from '../components/SplitSection';
-import StatsBox from '../components/StatsBox';
-import customerData from '../data/customer-data';
-import HeroImage from '../svg/HeroImage';
-import SvgCharts from '../svg/SvgCharts';
+import React, { Fragment } from 'react';
+import { Popover, Transition } from '@headlessui/react';
+import {
+  AnnotationIcon,
+  ChatAlt2Icon,
+  ChatAltIcon,
+  DocumentReportIcon,
+  HeartIcon,
+  EmojiHappyIcon,
+  ShoppingCartIcon,
+  MenuIcon,
+  PencilAltIcon,
+  QuestionMarkCircleIcon,
+  ReplyIcon,
+  SparklesIcon,
+  TrashIcon,
+  UsersIcon,
+  XIcon,
+} from '@heroicons/react/outline';
+import { ChevronDownIcon } from '@heroicons/react/solid';
 
-const Index = () => (
-  <Layout>
-    <section className="pt-20 md:pt-40">
-      <div className="container mx-auto px-8 lg:flex">
-        <div className="text-center lg:text-left lg:w-1/2">
-          <h1 className="text-4xl lg:text-5xl xl:text-6xl font-bold leading-none">
-            Main title of your landing page
-          </h1>
-          <p className="text-xl lg:text-2xl mt-6 font-light">
-            Free landing page template to promote your business startup and generate leads for the
-            offered services
-          </p>
-          <p className="mt-8 md:mt-12">
-            <Button size="lg">Get Started</Button>
-          </p>
-          <p className="mt-4 text-gray-600">Sed fermentum felis ut cursu</p>
-        </div>
-        <div className="lg:w-1/2">
-          <HeroImage />
-        </div>
-      </div>
-    </section>
-    <section id="features" className="py-20 lg:pb-40 lg:pt-48">
-      <div className="container mx-auto text-center">
-        <h2 className="text-3xl lg:text-5xl font-semibold">Main Features</h2>
-        <div className="flex flex-col sm:flex-row sm:-mx-3 mt-12">
-          <div className="flex-1 px-3">
-            <Card className="mb-8">
-              <p className="font-semibold text-xl">Service One</p>
-              <p className="mt-4">
-                An enim nullam tempor gravida donec enim ipsum blandit porta justo integer odio
-                velna vitae auctor integer.
-              </p>
-            </Card>
-          </div>
-          <div className="flex-1 px-3">
-            <Card className="mb-8">
-              <p className="font-semibold text-xl">Service Two</p>
-              <p className="mt-4">
-                An enim nullam tempor gravida donec enim ipsum blandit porta justo integer odio
-                velna vitae auctor integer.
-              </p>
-            </Card>
-          </div>
-          <div className="flex-1 px-3">
-            <Card className="mb-8">
-              <p className="font-semibold text-xl">Service Three</p>
-              <p className="mt-4">
-                An enim nullam tempor gravida donec enim ipsum blandit porta justo integer odio
-                velna vitae auctor integer.
-              </p>
-            </Card>
-          </div>
-        </div>
-      </div>
-    </section>
-    <SplitSection
-      id="services"
-      primarySlot={
-        <div className="lg:pr-32 xl:pr-48">
-          <h3 className="text-3xl font-semibold leading-tight">Market Analysis</h3>
-          <p className="mt-8 text-xl font-light leading-relaxed">
-            Our team of enthusiastic marketers will analyse and evaluate how your company stacks
-            against the closest competitors
-          </p>
-        </div>
-      }
-      secondarySlot={<SvgCharts />}
-    />
-    <SplitSection
-      reverseOrder
-      primarySlot={
-        <div className="lg:pl-32 xl:pl-48">
-          <h3 className="text-3xl font-semibold leading-tight">
-            Design And Plan Your Business Growth Steps
-          </h3>
-          <p className="mt-8 text-xl font-light leading-relaxed">
-            Once the market analysis process is completed our staff will search for opportunities
-            that are in reach
-          </p>
-        </div>
-      }
-      secondarySlot={<SvgCharts />}
-    />
-    <SplitSection
-      primarySlot={
-        <div className="lg:pr-32 xl:pr-48">
-          <h3 className="text-3xl font-semibold leading-tight">
-            Search For Performance Optimization
-          </h3>
-          <p className="mt-8 text-xl font-light leading-relaxed">
-            With all the information in place you will be presented with an action plan that your
-            company needs to follow
-          </p>
-        </div>
-      }
-      secondarySlot={<SvgCharts />}
-    />
-    <section id="stats" className="py-20 lg:pt-32">
-      <div className="container mx-auto text-center">
-        <LabelText className="text-gray-600">Our customers get results</LabelText>
-        <div className="flex flex-col sm:flex-row mt-8 lg:px-24">
-          <div className="w-full sm:w-1/3">
-            <StatsBox primaryText="+100%" secondaryText="Stats Information" />
-          </div>
-          <div className="w-full sm:w-1/3">
-            <StatsBox primaryText="+100%" secondaryText="Stats Information" />
-          </div>
-          <div className="w-full sm:w-1/3">
-            <StatsBox primaryText="+100%" secondaryText="Stats Information" />
-          </div>
-        </div>
-      </div>
-    </section>
-    <section id="testimonials" className="py-20 lg:py-40">
-      <div className="container mx-auto">
-        <LabelText className="mb-8 text-gray-600 text-center">What customers are saying</LabelText>
-        <div className="flex flex-col md:flex-row md:-mx-3">
-          {customerData.map(customer => (
-            <div key={customer.customerName} className="flex-1 px-3">
-              <CustomerCard customer={customer} />
+const solutions = [
+  {
+    name: 'Inbox',
+    description: 'Get a better understanding of where your traffic is coming from.',
+    href: '#',
+    icon: EmojiHappyIcon,
+  },
+  {
+    name: 'Messaging',
+    description: 'Speak directly to your customers in a more meaningful way.',
+    href: '#',
+    icon: AnnotationIcon,
+  },
+  {
+    name: 'Live Chat',
+    description: "Your customers' data will be safe and secure.",
+    href: '#',
+    icon: ChatAlt2Icon,
+  },
+  {
+    name: 'Knowledge Base',
+    description: "Connect with third-party tools that you're already using.",
+    href: '#',
+    icon: QuestionMarkCircleIcon,
+  },
+];
+const features = [
+  {
+    name: 'SURF',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+  {
+    name: 'FOIL',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+  {
+    name: 'SUP',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+  {
+    name: 'WETSUITS & JACKETS',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+  {
+    name: 'SHORTS',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+  {
+    name: 'SKATE',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+  {
+    name: 'CLOTHING',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+  {
+    name: 'WAKEBOARD',
+    description:
+      'Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.',
+    icon: ShoppingCartIcon,
+  },
+];
+const metrics = [
+  { id: 1, stat: '8K+', emphasis: 'Companies', rest: 'use laoreet amet lacus nibh integer quis.' },
+  { id: 2, stat: '25K+', emphasis: 'Countries around the globe', rest: 'lacus nibh integer quis.' },
+  {
+    id: 3,
+    stat: '98%',
+    emphasis: 'Customer satisfaction',
+    rest: 'laoreet amet lacus nibh integer quis.',
+  },
+  { id: 4, stat: '12M+', emphasis: 'Issues resolved', rest: 'lacus nibh integer quis.' },
+];
+const footerNavigation = {
+  shape: [
+    { name: 'Board shape', href: '#' },
+    { name: 'Prices', href: '#' },
+    { name: 'Wall of fame', href: '#' },
+    { name: 'Insights', href: '#' },
+  ],
+  school: [
+    { name: 'Lesson', href: '#' },
+    { name: 'Trip', href: '#' },
+    { name: 'Group', href: '#' },
+    { name: 'Book', href: '#' },
+  ],
+  shop: [
+    { name: 'Surf', href: '#' },
+    { name: 'Wetsuits', href: '#' },
+    { name: 'Skate', href: '#' },
+    { name: 'Clothing', href: '#' },
+    { name: 'Paddle', href: '#' },
+  ],
+  compagny: [
+    { name: 'About', href: '#' },
+    { name: 'Blog', href: '#' },
+    { name: 'Jobs', href: '#' },
+    { name: 'Press', href: '#' },
+    { name: 'Partners', href: '#' },
+  ],
+  social: [
+    {
+      name: 'Facebook',
+      href: '#',
+      icon: (props) => (
+        <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+          <path
+            fillRule="evenodd"
+            d="M22 12c0-5.523-4.477-10-10-10S2 6.477 2 12c0 4.991 3.657 9.128 8.438 9.878v-6.987h-2.54V12h2.54V9.797c0-2.506 1.492-3.89 3.777-3.89 1.094 0 2.238.195 2.238.195v2.46h-1.26c-1.243 0-1.63.771-1.63 1.562V12h2.773l-.443 2.89h-2.33v6.988C18.343 21.128 22 16.991 22 12z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+    {
+      name: 'Instagram',
+      href: '#',
+      icon: (props) => (
+        <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+          <path
+            fillRule="evenodd"
+            d="M12.315 2c2.43 0 2.784.013 3.808.06 1.064.049 1.791.218 2.427.465a4.902 4.902 0 011.772 1.153 4.902 4.902 0 011.153 1.772c.247.636.416 1.363.465 2.427.048 1.067.06 1.407.06 4.123v.08c0 2.643-.012 2.987-.06 4.043-.049 1.064-.218 1.791-.465 2.427a4.902 4.902 0 01-1.153 1.772 4.902 4.902 0 01-1.772 1.153c-.636.247-1.363.416-2.427.465-1.067.048-1.407.06-4.123.06h-.08c-2.643 0-2.987-.012-4.043-.06-1.064-.049-1.791-.218-2.427-.465a4.902 4.902 0 01-1.772-1.153 4.902 4.902 0 01-1.153-1.772c-.247-.636-.416-1.363-.465-2.427-.047-1.024-.06-1.379-.06-3.808v-.63c0-2.43.013-2.784.06-3.808.049-1.064.218-1.791.465-2.427a4.902 4.902 0 011.153-1.772A4.902 4.902 0 015.45 2.525c.636-.247 1.363-.416 2.427-.465C8.901 2.013 9.256 2 11.685 2h.63zm-.081 1.802h-.468c-2.456 0-2.784.011-3.807.058-.975.045-1.504.207-1.857.344-.467.182-.8.398-1.15.748-.35.35-.566.683-.748 1.15-.137.353-.3.882-.344 1.857-.047 1.023-.058 1.351-.058 3.807v.468c0 2.456.011 2.784.058 3.807.045.975.207 1.504.344 1.857.182.466.399.8.748 1.15.35.35.683.566 1.15.748.353.137.882.3 1.857.344 1.054.048 1.37.058 4.041.058h.08c2.597 0 2.917-.01 3.96-.058.976-.045 1.505-.207 1.858-.344.466-.182.8-.398 1.15-.748.35-.35.566-.683.748-1.15.137-.353.3-.882.344-1.857.048-1.055.058-1.37.058-4.041v-.08c0-2.597-.01-2.917-.058-3.96-.045-.976-.207-1.505-.344-1.858a3.097 3.097 0 00-.748-1.15 3.098 3.098 0 00-1.15-.748c-.353-.137-.882-.3-1.857-.344-1.023-.047-1.351-.058-3.807-.058zM12 6.865a5.135 5.135 0 110 10.27 5.135 5.135 0 010-10.27zm0 1.802a3.333 3.333 0 100 6.666 3.333 3.333 0 000-6.666zm5.338-3.205a1.2 1.2 0 110 2.4 1.2 1.2 0 010-2.4z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+    {
+      name: 'Twitter',
+      href: '#',
+      icon: (props) => (
+        <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+          <path d="M8.29 20.251c7.547 0 11.675-6.253 11.675-11.675 0-.178 0-.355-.012-.53A8.348 8.348 0 0022 5.92a8.19 8.19 0 01-2.357.646 4.118 4.118 0 001.804-2.27 8.224 8.224 0 01-2.605.996 4.107 4.107 0 00-6.993 3.743 11.65 11.65 0 01-8.457-4.287 4.106 4.106 0 001.27 5.477A4.072 4.072 0 012.8 9.713v.052a4.105 4.105 0 003.292 4.022 4.095 4.095 0 01-1.853.07 4.108 4.108 0 003.834 2.85A8.233 8.233 0 012 18.407a11.616 11.616 0 006.29 1.84" />
+        </svg>
+      ),
+    },
+    {
+      name: 'GitHub',
+      href: '#',
+      icon: (props) => (
+        <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+          <path
+            fillRule="evenodd"
+            d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+    {
+      name: 'Dribbble',
+      href: '#',
+      icon: (props) => (
+        <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+          <path
+            fillRule="evenodd"
+            d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10c5.51 0 10-4.48 10-10S17.51 2 12 2zm6.605 4.61a8.502 8.502 0 011.93 5.314c-.281-.054-3.101-.629-5.943-.271-.065-.141-.12-.293-.184-.445a25.416 25.416 0 00-.564-1.236c3.145-1.28 4.577-3.124 4.761-3.362zM12 3.475c2.17 0 4.154.813 5.662 2.148-.152.216-1.443 1.941-4.48 3.08-1.399-2.57-2.95-4.675-3.189-5A8.687 8.687 0 0112 3.475zm-3.633.803a53.896 53.896 0 013.167 4.935c-3.992 1.063-7.517 1.04-7.896 1.04a8.581 8.581 0 014.729-5.975zM3.453 12.01v-.26c.37.01 4.512.065 8.775-1.215.25.477.477.965.694 1.453-.109.033-.228.065-.336.098-4.404 1.42-6.747 5.303-6.942 5.629a8.522 8.522 0 01-2.19-5.705zM12 20.547a8.482 8.482 0 01-5.239-1.8c.152-.315 1.888-3.656 6.703-5.337.022-.01.033-.01.054-.022a35.318 35.318 0 011.823 6.475 8.4 8.4 0 01-3.341.684zm4.761-1.465c-.086-.52-.542-3.015-1.659-6.084 2.679-.423 5.022.271 5.314.369a8.468 8.468 0 01-3.655 5.715z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+  ],
+};
+
+function classNames(...classes) {
+  return classes.filter(Boolean).join(' ');
+}
+
+export default function Example() {
+  return (
+    <div className="bg-white">
+      <header>
+        <Popover className="relative bg-white">
+          {({ open }) => (
+            <>
+              <div className="flex justify-between items-center max-w-7xl mx-auto px-4 py-6 sm:px-6 md:justify-start md:space-x-10 lg:px-8">
+                <div className="flex justify-start lg:w-0 lg:flex-1">
+                  <a href="#">
+                    <span className="sr-only">Surf</span>
+                    {/* // added */}
+                    <div className="h-8  sm:h-10  flex justify-center items-center">
+                      <span className="whitespace-nowrap text-base font-medium text-gray-500 hover:text-gray-900">
+                        {`<logo rod's surf>`}
+                      </span>
+                    </div>
+                    {/* <img
+                      className="h-8 w-auto sm:h-10"
+                      src="https://tailwindui.com/img/logos/workflow-mark-purple-600-to-indigo-600.svg"
+                      alt=""
+                    /> */}
+                  </a>
+                </div>
+                <div className="-mr-2 -my-2 md:hidden">
+                  <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
+                    <span className="sr-only">Open menu</span>
+                    <MenuIcon className="h-6 w-6" aria-hidden="true" />
+                  </Popover.Button>
+                </div>
+                <Popover.Group as="nav" className="hidden md:flex space-x-10">
+                  {/* <Popover className="relative">
+                    {({ open }) => (
+                      <>
+                        <Popover.Button
+                          className={classNames(
+                            open ? 'text-gray-900' : 'text-gray-500',
+                            'group bg-white rounded-md inline-flex items-center text-base font-medium hover:text-gray-900 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500'
+                          )}
+                        >
+                          <span>Solutions</span>
+                          <ChevronDownIcon
+                            className={classNames(
+                              open ? 'text-gray-600' : 'text-gray-400',
+                              'ml-2 h-5 w-5 group-hover:text-gray-500'
+                            )}
+                            aria-hidden="true"
+                          />
+                        </Popover.Button>
+
+                        <Transition
+                          show={open}
+                          as={Fragment}
+                          enter="transition ease-out duration-200"
+                          enterFrom="opacity-0 translate-y-1"
+                          enterTo="opacity-100 translate-y-0"
+                          leave="transition ease-in duration-150"
+                          leaveFrom="opacity-100 translate-y-0"
+                          leaveTo="opacity-0 translate-y-1"
+                        >
+                          <Popover.Panel
+                            static
+                            className="absolute z-10 -ml-4 mt-3 transform w-screen max-w-md lg:max-w-2xl lg:ml-0 lg:left-1/2 lg:-translate-x-1/2"
+                          >
+                            <div className="rounded-lg shadow-lg ring-1 ring-black ring-opacity-5 overflow-hidden">
+                              <div className="relative grid gap-6 bg-white px-5 py-6 sm:gap-8 sm:p-8 lg:grid-cols-2">
+                                {solutions.map((item) => (
+                                  <a
+                                    key={item.name}
+                                    href={item.href}
+                                    className="-m-3 p-3 flex items-start rounded-lg hover:bg-gray-50"
+                                  >
+                                    <div className="flex-shrink-0 flex items-center justify-center h-10 w-10 rounded-md bg-gradient-to-r from-purple-600 to-indigo-600 text-white sm:h-12 sm:w-12">
+                                      <item.icon className="h-6 w-6" aria-hidden="true" />
+                                    </div>
+                                    <div className="ml-4">
+                                      <p className="text-base font-medium text-gray-900">
+                                        {item.name}
+                                      </p>
+                                      <p className="mt-1 text-sm text-gray-500">
+                                        {item.description}
+                                      </p>
+                                    </div>
+                                  </a>
+                                ))}
+                              </div>
+                            </div>
+                          </Popover.Panel>
+                        </Transition>
+                      </>
+                    )}
+                  </Popover> */}
+
+                  <a href="#" className="text-base font-medium text-gray-500 hover:text-gray-900">
+                    Shape
+                  </a>
+                  <a href="#" className="text-base font-medium text-gray-500 hover:text-gray-900">
+                    School
+                  </a>
+                  <a href="#" className="text-base font-medium text-gray-500 hover:text-gray-900">
+                    Shop
+                  </a>
+                </Popover.Group>
+                {/* <div className="hidden md:flex items-center justify-end md:flex-1 lg:w-0">
+                  <a
+                    href="#"
+                    className="whitespace-nowrap text-base font-medium text-gray-500 hover:text-gray-900"
+                  >
+                    Sign in
+                  </a>
+                  <a
+                    href="#"
+                    className="ml-8 whitespace-nowrap inline-flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700"
+                  >
+                    Sign up
+                  </a>
+                </div> */}
+              </div>
+
+              <Transition
+                show={open}
+                as={Fragment}
+                enter="duration-200 ease-out"
+                enterFrom="opacity-0 scale-95"
+                enterTo="opacity-100 scale-100"
+                leave="duration-100 ease-in"
+                leaveFrom="opacity-100 scale-100"
+                leaveTo="opacity-0 scale-95"
+              >
+                <Popover.Panel
+                  focus
+                  static
+                  className="absolute z-30 top-0 inset-x-0 p-2 transition transform origin-top-right md:hidden"
+                >
+                  <div className="rounded-lg shadow-lg ring-1 ring-black ring-opacity-5 bg-white divide-y-2 divide-gray-50">
+                    <div className="pt-5 pb-6 px-5">
+                      <div className="flex items-center justify-between">
+                        <div>
+                          <img
+                            className="h-8 w-auto"
+                            src="https://tailwindui.com/img/logos/workflow-mark-purple-600-to-indigo-600.svg"
+                            alt="Workflow"
+                          />
+                        </div>
+                        <div className="-mr-2">
+                          <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
+                            <span className="sr-only">Close menu</span>
+                            <XIcon className="h-6 w-6" aria-hidden="true" />
+                          </Popover.Button>
+                        </div>
+                      </div>
+                      <div className="mt-6">
+                        {/* <nav className="grid grid-cols-1 gap-7">
+                          {solutions.map((item) => (
+                            <a
+                              key={item.name}
+                              href={item.href}
+                              className="-m-3 p-3 flex items-center rounded-lg hover:bg-gray-50"
+                            >
+                              <div className="flex-shrink-0 flex items-center justify-center h-10 w-10 rounded-md bg-gradient-to-r from-purple-600 to-indigo-600 text-white">
+                                <item.icon className="h-6 w-6" aria-hidden="true" />
+                              </div>
+                              <div className="ml-4 text-base font-medium text-gray-900">
+                                {item.name}
+                              </div>
+                            </a>
+                          ))}
+                        </nav> */}
+                      </div>
+                    </div>
+                    <div className="py-6 px-5">
+                      <div className="grid grid-cols-2 gap-4">
+                        <a
+                          href="#"
+                          className="text-base font-medium text-gray-900 hover:text-gray-700"
+                        >
+                          Pricing
+                        </a>
+                        <a
+                          href="#"
+                          className="text-base font-medium text-gray-900 hover:text-gray-700"
+                        >
+                          Partners
+                        </a>
+                        <a
+                          href="#"
+                          className="text-base font-medium text-gray-900 hover:text-gray-700"
+                        >
+                          Company
+                        </a>
+                      </div>
+                      <div className="mt-6">
+                        <a
+                          href="#"
+                          className="w-full flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700"
+                        >
+                          Sign up
+                        </a>
+                        <p className="mt-6 text-center text-base font-medium text-gray-500">
+                          Existing customer?
+                          <a href="#" className="text-gray-900">
+                            Sign in
+                          </a>
+                        </p>
+                      </div>
+                    </div>
+                  </div>
+                </Popover.Panel>
+              </Transition>
+            </>
+          )}
+        </Popover>
+      </header>
+
+      <main>
+        {/* Hero section */}
+        <div className="relative">
+          <div className="absolute inset-x-0 bottom-0 h-1/2 bg-gray-100" />
+          <div className="max-w-7xl mx-auto sm:px-6 lg:px-8">
+            <div className="relative shadow-xl sm:rounded-2xl sm:overflow-hidden">
+              <div className="absolute inset-0">
+                <img
+                  className="h-full w-full object-cover"
+                  src="https://images.unsplash.com/photo-1475706997440-9f2a24435b83?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=3689&q=80&ixlib=rb-1.2.1&auto=format&fit=crop&w=2830&q=80" //&sat=-100"
+                  alt="People working on laptops"
+                />
+                {/* // DON'T WORK */}
+                {/* <div className="absolute inset-0 bg-indigo-700 mix-blend-multiply" /> */}
+              </div>
+              <div className="relative px-4 py-16 sm:px-6 sm:py-24 lg:py-32 lg:px-8">
+                <h1 className="text-center text-4xl font-extrabold tracking-tight sm:text-5xl lg:text-6xl">
+                  <span className="block text-white">Rod's surf</span>
+                  {/* <span className="block text-indigo-200">customer support</span> */}
+                </h1>
+                <p className="mt-6 max-w-lg mx-auto text-center text-xl text-white sm:max-w-3xl">
+                  Learn to surf, equip yourself or shape your own surfboard at Rod's surf carcans
+                </p>
+                <div className="mt-10 max-w-sm mx-auto sm:max-w-none sm:flex sm:justify-center">
+                  <div className="space-y-4 sm:space-y-0 sm:mx-auto sm:inline-grid sm:grid-cols-2 sm:gap-5">
+                    <a
+                      href="#"
+                      className="flex items-center justify-center px-4 py-3 border border-transparent text-base font-medium rounded-md shadow-sm text-indigo-700 bg-white hover:bg-indigo-50 sm:px-8"
+                    >
+                      Book online
+                    </a>
+                    <a
+                      href="#"
+                      className="flex items-center justify-center px-4 py-3 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-indigo-500 bg-opacity-60 hover:bg-opacity-70 sm:px-8"
+                    >
+                      Shape a board
+                    </a>
+                  </div>
+                </div>
+              </div>
             </div>
-          ))}
+          </div>
         </div>
-      </div>
-    </section>
-    <section className="container mx-auto my-20 py-24 bg-gray-200 rounded-lg text-center">
-      <h3 className="text-5xl font-semibold">Ready to grow your business?</h3>
-      <p className="mt-8 text-xl font-light">
-        Quis lectus nulla at volutpat diam ut. Enim lobortis scelerisque fermentum dui faucibus in.
-      </p>
-      <p className="mt-8">
-        <Button size="xl">Get Started Now</Button>
-      </p>
-    </section>
-  </Layout>
-);
 
-export default Index;
+        {/* Logo Cloud */}
+        {/* <div className="bg-gray-100">
+          <div className="max-w-7xl mx-auto py-16 px-4 sm:px-6 lg:px-8">
+            <p className="text-center text-sm font-semibold uppercase text-gray-500 tracking-wide">
+              Trusted by over 5 very average small businesses
+            </p>
+            <div className="mt-6 grid grid-cols-2 gap-8 md:grid-cols-6 lg:grid-cols-5">
+              <div className="col-span-1 flex justify-center md:col-span-2 lg:col-span-1">
+                <img
+                  className="h-12"
+                  src="https://tailwindui.com/img/logos/tuple-logo-gray-400.svg"
+                  alt="Tuple"
+                />
+              </div>
+              <div className="col-span-1 flex justify-center md:col-span-2 lg:col-span-1">
+                <img
+                  className="h-12"
+                  src="https://tailwindui.com/img/logos/mirage-logo-gray-400.svg"
+                  alt="Mirage"
+                />
+              </div>
+              <div className="col-span-1 flex justify-center md:col-span-2 lg:col-span-1">
+                <img
+                  className="h-12"
+                  src="https://tailwindui.com/img/logos/statickit-logo-gray-400.svg"
+                  alt="StaticKit"
+                />
+              </div>
+              <div className="col-span-1 flex justify-center md:col-span-2 md:col-start-2 lg:col-span-1">
+                <img
+                  className="h-12"
+                  src="https://tailwindui.com/img/logos/transistor-logo-gray-400.svg"
+                  alt="Transistor"
+                />
+              </div>
+              <div className="col-span-2 flex justify-center md:col-span-2 md:col-start-4 lg:col-span-1">
+                <img
+                  className="h-12"
+                  src="https://tailwindui.com/img/logos/workcation-logo-gray-400.svg"
+                  alt="Workcation"
+                />
+              </div>
+            </div>
+          </div>
+        </div> */}
+
+        {/* Alternating Feature Sections */}
+        <div className="relative pt-16 pb-32 overflow-hidden">
+          <div
+            aria-hidden="true"
+            className="absolute inset-x-0 top-0 h-48 bg-gradient-to-b from-gray-100"
+          />
+          <div className="relative">
+            <div className="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
+              <div className="px-4 max-w-xl mx-auto sm:px-6 lg:py-16 lg:max-w-none lg:mx-0 lg:px-0">
+                <div>
+                  <div>
+                    <span className="h-12 w-12 rounded-md flex items-center justify-center bg-gradient-to-r from-purple-600 to-indigo-600">
+                      <EmojiHappyIcon className="h-6 w-6 text-white" aria-hidden="true" />
+                    </span>
+                  </div>
+                  <div className="mt-6">
+                    <h2 className="text-3xl font-extrabold tracking-tight text-gray-900">
+                      Come live a new experience, share sensations and create new memories
+                    </h2>
+                    <p className="mt-4 text-lg text-gray-500">
+                      Semper curabitur ullamcorper posuere nunc sed. Ornare iaculis bibendum
+                      malesuada faucibus lacinia porttitor. Pulvinar laoreet sagittis viverra duis.
+                      In venenatis sem arcu pretium pharetra at. Lectus viverra dui tellus ornare
+                      pharetra.
+                    </p>
+                    <div className="mt-6">
+                      <a
+                        href="#"
+                        className="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700"
+                      >
+                        Learn to surf
+                      </a>
+                    </div>
+                  </div>
+                </div>
+                <div className="mt-8 border-t border-gray-200 pt-6">
+                  <blockquote>
+                    <div>
+                      <p className="text-base text-gray-500">
+                        &ldquo;Cras velit quis eros eget rhoncus lacus ultrices sed diam. Sit orci
+                        risus aenean curabitur donec aliquet. Mi venenatis in euismod ut.&rdquo;
+                      </p>
+                    </div>
+                    <footer className="mt-3">
+                      <div className="flex items-center space-x-3">
+                        <div className="flex-shrink-0">
+                          <img
+                            className="h-6 w-6 rounded-full"
+                            src="https://images.unsplash.com/photo-1509783236416-c9ad59bae472?ixlib=rb-=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=8&w=1024&h=1024&q=80"
+                            alt=""
+                          />
+                        </div>
+                        <div className="text-base font-medium text-gray-700">
+                          Marcia Hill and her family
+                        </div>
+                      </div>
+                    </footer>
+                  </blockquote>
+                </div>
+              </div>
+              <div className="mt-12 sm:mt-16 lg:mt-0">
+                <div className="pl-4 -mr-48 sm:pl-6 md:-mr-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
+                  <img
+                    className="w-full rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:left-0 lg:h-full lg:w-auto lg:max-w-none"
+                    src="https://images.unsplash.com/photo-1495819427834-1954f20ebb97?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=3750&q=80"
+                    alt="surfboards"
+                  />
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className="mt-24">
+            <div className="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
+              <div className="px-4 max-w-xl mx-auto sm:px-6 lg:py-32 lg:max-w-none lg:mx-0 lg:px-0 lg:col-start-2">
+                <div>
+                  <div>
+                    <span className="h-12 w-12 rounded-md flex items-center justify-center bg-gradient-to-r from-purple-600 to-indigo-600">
+                      <SparklesIcon className="h-6 w-6 text-white" aria-hidden="true" />
+                    </span>
+                  </div>
+                  <div className="mt-6">
+                    <h2 className="text-3xl font-extrabold tracking-tight text-gray-900">
+                      Shape your own board
+                    </h2>
+                    <p className="mt-4 text-lg text-gray-500">
+                      Semper curabitur ullamcorper posuere nunc sed. Ornare iaculis bibendum
+                      malesuada faucibus lacinia porttitor. Pulvinar laoreet sagittis viverra duis.
+                      In venenatis sem arcu pretium pharetra at. Lectus viverra dui tellus ornare
+                      pharetra.
+                    </p>
+                    <div className="mt-6">
+                      <a
+                        href="#"
+                        className="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700"
+                      >
+                        Discover our workshop
+                      </a>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div className="mt-12 sm:mt-16 lg:mt-0 lg:col-start-1">
+                <div className="pr-4 -ml-48 sm:pr-6 md:-ml-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
+                  <img
+                    className="w-full rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:right-0 lg:h-full lg:w-auto lg:max-w-none"
+                    src="https://images.unsplash.com/photo-1537519646099-335112f03225?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=3750&q=80"
+                    alt="Surf workshop"
+                  />
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        {/* Gradient Feature Section */}
+        <div className="bg-gradient-to-r from-purple-800 to-indigo-700">
+          <div className="max-w-4xl mx-auto px-4 py-16 sm:px-6 sm:pt-20 sm:pb-24 lg:max-w-7xl lg:pt-24 lg:px-8">
+            <h2 className="text-3xl font-extrabold text-white tracking-tight">
+              Find all the gear you need
+            </h2>
+            <p className="mt-4 max-w-3xl text-lg text-purple-200">
+              Specialist in board sports for 38 years, Rod's offers you the best brands of
+              boardsports:
+            </p>
+            <div className="mt-12 grid grid-cols-1 gap-x-6 gap-y-12 sm:grid-cols-2 lg:mt-16 lg:grid-cols-4 lg:gap-x-8 lg:gap-y-16">
+              {features.map((feature) => (
+                <div key={feature.name}>
+                  <div>
+                    <span className="flex items-center justify-center h-12 w-12 rounded-md bg-white bg-opacity-10">
+                      <feature.icon className="h-6 w-6 text-indigo-700" aria-hidden="true" />
+                    </span>
+                  </div>
+                  <div className="mt-6">
+                    <h3 className="text-lg font-medium text-white">{feature.name}</h3>
+                    <p className="mt-2 text-base text-purple-200">{feature.description}</p>
+                  </div>
+                </div>
+              ))}
+            </div>
+          </div>
+        </div>
+
+        {/* Stats section */}
+        <div className="relative bg-gray-900">
+          <div className="h-80 absolute inset-x-0 bottom-0 xl:top-0 xl:h-full">
+            <div className="h-full w-full xl:grid xl:grid-cols-2">
+              <div className="h-full xl:relative xl:col-start-2">
+                <img
+                  className="h-full w-full object-cover opacity-25 xl:absolute xl:inset-0"
+                  src="https://images.unsplash.com/photo-1575623811814-dfbb2c13e4ad?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=3750&q=80&sat=-100"
+                  alt="People working on laptops"
+                />
+                <div
+                  aria-hidden="true"
+                  className="absolute inset-x-0 top-0 h-32 bg-gradient-to-b from-gray-900 xl:inset-y-0 xl:left-0 xl:h-full xl:w-32 xl:bg-gradient-to-r"
+                />
+              </div>
+            </div>
+          </div>
+          <div className="max-w-4xl mx-auto px-4 sm:px-6 lg:max-w-7xl lg:px-8 xl:grid xl:grid-cols-2 xl:grid-flow-col-dense xl:gap-x-8">
+            <div className="relative pt-12 pb-64 sm:pt-24 sm:pb-64 xl:col-start-1 xl:pb-24">
+              <h2 className="text-sm font-semibold tracking-wide uppercase">
+                <span className="bg-gradient-to-r from-purple-300 to-indigo-300 bg-clip-text text-transparent">
+                  Rod's restaurant
+                </span>
+              </h2>
+              <p className="mt-3 text-3xl font-extrabold text-white">
+                Treat yourself after a surf session
+              </p>
+              <p className="mt-5 text-lg text-gray-300">
+                Rhoncus sagittis risus arcu erat lectus bibendum. Ut in adipiscing quis in viverra
+                tristique sem. Ornare feugiat viverra eleifend fusce orci in quis amet. Sit in et
+                vitae tortor, massa. Dapibus laoreet amet lacus nibh integer quis. Eu vulputate diam
+                sit tellus quis at.
+              </p>
+              {/* <div className="mt-12 grid grid-cols-1 gap-y-12 gap-x-6 sm:grid-cols-2">
+                {metrics.map((item) => (
+                  <p key={item.id}>
+                    <span className="block text-2xl font-bold text-white">{item.stat}</span>
+                    <span className="mt-1 block text-base text-gray-300">
+                      <span className="font-medium text-white">{item.emphasis}</span> {item.rest}
+                    </span>
+                  </p>
+                ))}
+              </div> */}
+            </div>
+          </div>
+        </div>
+
+        {/* CTA Section */}
+        <div className="bg-white">
+          <div className="max-w-4xl mx-auto py-16 px-4 sm:px-6 sm:py-24 lg:max-w-7xl lg:px-8 lg:flex lg:items-center lg:justify-between">
+            <h2 className="text-4xl font-extrabold tracking-tight text-gray-900 sm:text-4xl">
+              <span className="block">Ready to get started?</span>
+              <span className="block bg-gradient-to-r from-purple-600 to-indigo-600 bg-clip-text text-transparent">
+                Get in touch
+              </span>
+            </h2>
+            <div className="mt-6 space-y-4 sm:space-y-0 sm:flex sm:space-x-5 flex">
+              <a
+                href="#"
+                className="flex items-center justify-center px-4 py-3 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700"
+              >
+                Learn more
+              </a>
+              <a
+                href="#"
+                className="flex items-center justify-center px-4 py-3 border border-transparent text-base font-medium rounded-md shadow-sm text-indigo-800 bg-indigo-50 hover:bg-indigo-100"
+              >
+                Get started
+              </a>
+            </div>
+          </div>
+        </div>
+      </main>
+
+      <footer className="bg-gray-50" aria-labelledby="footerHeading">
+        <h2 id="footerHeading" className="sr-only">
+          Footer
+        </h2>
+        <div className="max-w-7xl mx-auto pt-16 pb-8 px-4 sm:px-6 lg:pt-24 lg:px-8">
+          <div className="xl:grid xl:grid-cols-3 xl:gap-8">
+            <div className="grid grid-cols-2 gap-8 xl:col-span-2">
+              <div className="md:grid md:grid-cols-2 md:gap-8">
+                <div>
+                  <h3 className="text-sm font-semibold text-gray-400 tracking-wider uppercase">
+                    Shape
+                  </h3>
+                  <ul className="mt-4 space-y-4">
+                    {footerNavigation.shape.map((item) => (
+                      <li key={item.name}>
+                        <a href={item.href} className="text-base text-gray-500 hover:text-gray-900">
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+                <div className="mt-12 md:mt-0">
+                  <h3 className="text-sm font-semibold text-gray-400 tracking-wider uppercase">
+                    School
+                  </h3>
+                  <ul className="mt-4 space-y-4">
+                    {footerNavigation.school.map((item) => (
+                      <li key={item.name}>
+                        <a href={item.href} className="text-base text-gray-500 hover:text-gray-900">
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+              </div>
+              <div className="md:grid md:grid-cols-2 md:gap-8">
+                <div>
+                  <h3 className="text-sm font-semibold text-gray-400 tracking-wider uppercase">
+                    Shop
+                  </h3>
+                  <ul className="mt-4 space-y-4">
+                    {footerNavigation.shop.map((item) => (
+                      <li key={item.name}>
+                        <a href={item.href} className="text-base text-gray-500 hover:text-gray-900">
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+                <div className="mt-12 md:mt-0">
+                  <h3 className="text-sm font-semibold text-gray-400 tracking-wider uppercase">
+                    Compagny
+                  </h3>
+                  <ul className="mt-4 space-y-4">
+                    {footerNavigation.compagny.map((item) => (
+                      <li key={item.name}>
+                        <a href={item.href} className="text-base text-gray-500 hover:text-gray-900">
+                          {item.name}
+                        </a>
+                      </li>
+                    ))}
+                  </ul>
+                </div>
+              </div>
+            </div>
+            <div className="mt-12 xl:mt-0">
+              <h3 className="text-sm font-semibold text-gray-400 tracking-wider uppercase">
+                Subscribe to our newsletter
+              </h3>
+              <p className="mt-4 text-base text-gray-500">
+                The latest news, articles, and resources, sent to your inbox weekly.
+              </p>
+              <form className="mt-4 sm:flex sm:max-w-md">
+                <label htmlFor="emailAddress" className="sr-only">
+                  Email address
+                </label>
+                <input
+                  type="email"
+                  name="emailAddress"
+                  id="emailAddress"
+                  autoComplete="email"
+                  required
+                  className="appearance-none min-w-0 w-full bg-white border border-gray-300 rounded-md shadow-sm py-2 px-4 text-base text-gray-900 placeholder-gray-500 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:placeholder-gray-400"
+                  placeholder="Enter your email"
+                />
+                <div className="mt-3 rounded-md sm:mt-0 sm:ml-3 sm:flex-shrink-0">
+                  <button
+                    type="submit"
+                    className="w-full flex items-center justify-center px-4 py-3 border border-transparent text-base font-medium rounded-md shadow-sm text-white bg-gradient-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700"
+                  >
+                    Subscribe
+                  </button>
+                </div>
+              </form>
+            </div>
+          </div>
+          <div className="mt-12 border-t border-gray-200 pt-8 md:flex md:items-center md:justify-between lg:mt-16">
+            <div className="flex space-x-6 md:order-2">
+              {footerNavigation.social.map((item) => (
+                <a key={item.name} href={item.href} className="text-gray-400 hover:text-gray-500">
+                  <span className="sr-only">{item.name}</span>
+                  <item.icon className="h-6 w-6" aria-hidden="true" />
+                </a>
+              ))}
+            </div>
+            <p className="mt-8 text-base text-gray-400 md:mt-0 md:order-1">
+              &copy; 2020 Workflow, Inc. All rights reserved.
+            </p>
+          </div>
+        </div>
+      </footer>
+    </div>
+  );
+}
